query/requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain-community
pydantic-settings
sentence-transformers
faiss-cpu
pinecone-client